AliveCheck sends notification only if it is in its configuration. When creating new task instance, catch if storing of configuration to database was unsuccessful.

+++ framework2/tasks/AliveCheck/src/AliveCheckManager.cpp
@@ -176,7 +176,7 @@ AliveCheckConfig AliveCheckManager::parseConfiguration(std::map<std::string, std
         throw std::runtime_error("Could not parse received configuration.");
     }
     else {
-        parsed_config.send_notif = std::stoll(send_notif_it->second);
+        parsed_config.send_notif = std::stoi(send_notif_it->second);
     }
     return parsed_config;
 }
